fix(mon): add acceptable mongo errors in utils and avoid error logs for them
