+++ core/stores/mon/collection.go
@@ -3,15 +3,12 @@ package mon
 
 import (
 	"context"
-	"errors"
 	"time"
 
 	"github.com/zeromicro/go-zero/core/breaker"
-	"github.com/zeromicro/go-zero/core/errorx"
 	"github.com/zeromicro/go-zero/core/timex"
 	"go.mongodb.org/mongo-driver/v2/mongo"
 	"go.mongodb.org/mongo-driver/v2/mongo/options"
-	"go.mongodb.org/mongo-driver/v2/x/mongo/driver/session"
 )
 
 const (
@@ -549,25 +546,6 @@ func (p keepablePromise) keep(err error) error {
 
 	p.log(err)
 	return err
-}
-
-func acceptable(err error) bool {
-	return err == nil || isDupKeyError(err) ||
-		errorx.In(err, mongo.ErrNoDocuments, mongo.ErrNilValue,
-			mongo.ErrNilDocument, mongo.ErrNilCursor, mongo.ErrEmptySlice,
-			// session errors
-			session.ErrSessionEnded, session.ErrNoTransactStarted, session.ErrTransactInProgress,
-			session.ErrAbortAfterCommit, session.ErrAbortTwice, session.ErrCommitAfterAbort,
-			session.ErrUnackWCUnsupported, session.ErrSnapshotTransaction)
-}
-
-func isDupKeyError(err error) bool {
-	var e mongo.WriteException
-	if !errors.As(err, &e) {
-		return false
-	}
-
-	return e.HasErrorCode(duplicateKeyCode)
 }
 
 // monCollection defines a MongoDB collection, used for unit test
