+++ core/stores/mon/collection_test.go
@@ -445,48 +445,6 @@ func TestDecoratedCollection_LogDuration(t *testing.T) {
 	assert.Contains(t, buf.String(), "slowcall")
 }
 
-func TestAcceptable(t *testing.T) {
-	tests := []struct {
-		name string
-		err  error
-		want bool
-	}{
-		{"NilError", nil, true},
-		{"NoDocuments", mongo.ErrNoDocuments, true},
-		{"NilValue", mongo.ErrNilValue, true},
-		{"NilDocument", mongo.ErrNilDocument, true},
-		{"NilCursor", mongo.ErrNilCursor, true},
-		{"EmptySlice", mongo.ErrEmptySlice, true},
-		{"DuplicateKeyError", mongo.WriteException{WriteErrors: []mongo.WriteError{{Code: duplicateKeyCode}}}, true},
-		{"OtherError", errors.New("other error"), false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, acceptable(tt.err))
-		})
-	}
-}
-
-func TestIsDupKeyError(t *testing.T) {
-	tests := []struct {
-		name string
-		err  error
-		want bool
-	}{
-		{"NilError", nil, false},
-		{"NonDupKeyError", errors.New("some other error"), false},
-		{"DupKeyError", mongo.WriteException{WriteErrors: []mongo.WriteError{{Code: duplicateKeyCode}}}, true},
-		{"OtherMongoError", mongo.WriteException{WriteErrors: []mongo.WriteError{{Code: 12345}}}, false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, isDupKeyError(tt.err))
-		})
-	}
-}
-
 func newTestCollection(collection monCollection, brk breaker.Breaker) *decoratedCollection {
 	return &decoratedCollection{
 		Collection: collection,
